requirements.txt
# LyricLawyer Requirements

# Core framework
google-adk>=1.8.0
google-generativeai>=0.3.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Data processing
numpy>=1.24.0
python-dotenv>=1.0.0

# Database
# (SQLite is built into Python)

# Optional: For enhanced features
# requests>=2.31.0  # For future API integrations